Remove singleton and replace with patch
