+++ Commands/OpenLogWindow.cs
@@ -23,7 +23,12 @@ public class OpenLogWindow : AbstractConsoleCmd
 
     public static void OpenWindow(bool stealFocus)
     {
-        if (!BaseLibMain.IsMainThread) throw new InvalidOperationException("OpenWindow called when not on main thread");
+        if (!BaseLibMain.IsMainThread)
+        {
+            BaseLibMain.Logger.Info("OpenWindow called when not on main thread");
+            return;
+        }
+        
         var instance = NGame.Instance;
         if (instance == null) return;
